fix: add authorization checks for product management actions in ProductController

+++ app/Http/Controllers/ProductController.php
@@ -7,11 +7,15 @@ use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
+
     public function index()
     {
+        //  req gates for view  Products
+        Gate::authorize('view', Product::class);
         $products = Product::paginate(10);
         return ProductResource::collection($products);
    }
@@ -20,6 +24,8 @@ class ProductController extends Controller
      */
     public function show(string $id)
     {
+        //  req gates for view  one single Products
+        Gate::authorize('view', Product::class);
         $product = Product::findOrFail($id);
         return new ProductResource($product);
     }
@@ -30,6 +36,8 @@ class ProductController extends Controller
      */
     public function store(ProductCreateRequest $request)
     {
+        //  req gates for add  Products
+        Gate::authorize('edit', Product::class);
         $product = Product::create($request->only('title', 'description','image', 'price') );
         return response(new ProductResource($product), Response::HTTP_CREATED);
     }
@@ -37,6 +45,8 @@ class ProductController extends Controller
 
     public function update(ProductCreateRequest $request, string $id)
     {
+        //  req gates for update  Products
+        Gate::authorize('edit', Product::class);
         $product = Product::findOrFail($id);
         $product->update($request->only('title', 'description', 'image', 'price'));
         return new ProductResource($product);
@@ -47,6 +57,8 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+        //  req gates for delete  Products
+        Gate::authorize('edit', Product::class);
         Product::findOrFail($id)->delete();
         return response()->json(['message' => 'Product deleted successfully'], Response::HTTP_OK);
     }
